Add files via upload
The CONTENT HOME file is a content home page like we discussed in the meeting on 11/20. The home page I made last week can stay the same, this one should just become a landing page inbetweeen the home page and the content pages.

The other files are for the slideshow. If it's easier for me to update the html files to get these working I'm more than happy to, otherwise let me know if you need help! The three swipe files are for the slideshow you can go through manually (for content page(s)), and the auto is the one that goes through on its own (for the homepage)